src/rf/requirements/base.txt
Django>=1.8,<1.8.99
gunicorn==19.1.1
django-statsd-mozilla==0.3.15
django-redis==3.8.3
hiredis==0.1.6
djangorestframework==3.1.1
django-registration-redux==1.2
boto==2.34.0
